+++ src/index.ts
@@ -2,10 +2,7 @@ import {log} from './utils/log.js';
 import express, {ErrorRequestHandler} from 'express';
 import {StatusCodes} from 'http-status-codes';
 import {getEnv} from './utils/getEnv.js';
-import {deleteUserRoute} from './routes/user/deleteUserRoute.js';
-import {getUserRoute} from './routes/user/getUserRoute.js';
-import {postUserRoute} from './routes/user/postUserRoute.js';
-import {patchUserRoute} from './routes/user/patchUserRoute.js';
+import {routes} from './routes/index.js';
 
 const NODE_ENV = getEnv('NODE_ENV');
 const HOST = getEnv('HOST');
@@ -31,10 +28,7 @@ if (NODE_ENV === 'development') {
   server.use('/_coverage', express.static('build/coverage/lcov-report'));
 }
 
-server.use(postUserRoute);
-server.use(getUserRoute);
-server.use(patchUserRoute);
-server.use(deleteUserRoute);
+server.use(routes);
 
 /**
  * Error handling.
